requirements.txt
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
oauth2client==4.1.3
packaging==21.3
pbr==5.8.1
pluggy==1.0.0
protobuf==3.20.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==3.0.7
pytest==7.1.1
pytest-html==3.1.1
pytest-metadata==2.0.1
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
testresources==2.0.1
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
